% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zfa.example}
\alias{zfa.example}
\title{Example data for zfa}
\format{The \code{zfa.example} contains the following objects:

\code{X}

a numeric genotype matrix of 1800 individuals and 512 rare variants. Each row represents a different individual, and each column represents a different rare genetic variant. Genotypes are coded as 0,1 or 2.

\code{y}

a numeric vector of binary phenotypes, among which there are 343 cases and 1457 controls.}
\usage{
data(zfa.example)
}
\description{
Example data for \code{"zfa"} function.
}
\seealso{
\code{\link{zfa}}
}
\keyword{datasets}

